Update 0-rectangle.py
check the format using cmd: pycodestyle file.py
The PEP 8 style checker (pycodestyle) is indicating a warning (W391) about having a blank line at the end of the file. According to PEP 8, Python files should generally not end with a blank line.

To resolve this warning, you can simply remove the blank line at the end of the file so that it ends after the last line of code. Here's the corrected version of the code